2-demon-of-gigantism-tiers-and-models/pyproject.toml
[tool.poetry]
name = "2-demon-of-gigantism-tiers-and-models"
version = "0.1.0"
description = "Demo for the Demon of Gigantism."
authors = ["Eugene Klyuchnikov <eugene.klyuchnikov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"  # Specifies Python 3.11.x is required

# Core dependencies for all plays
pyyaml = "^6.0.3"
rich = "^14.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"